requirements.txt
babel==2.16.0
bidict==0.23.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
et-xmlfile==1.1.0
h11==0.14.0
idna==3.10
numpy==2.1.1
openpyxl==3.1.5
pandas==2.2.3
pillow==10.4.0
pyodbc==5.2.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.9.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2024.2
reportlab==4.2.5
requests==2.32.3
simple-websocket==1.1.0
six==1.16.0
tabulate==0.9.0
tkcalendar==1.6.1
tzdata==2024.2
urllib3==2.2.3
wsproto==1.2.0